Update README.md
@castleyeo @RobertLRead 

I updated this README with the information from the project page, calling for volunteers with specific skillsets. I also added a blurb at the end that might be helpful to add to the end of each README. Please commit the changes unless you have any corrections, and I'll continue moving through the projects to update them.